requirements.txt
# Python 의존성

# Google Gemini API
google-generativeai>=0.8.3

# RSS 파싱
feedparser>=6.0.11

# HTTP 요청
requests>=2.31.0

# 환경 변수 관리
python-dotenv>=1.0.0
